backend/requirements.txt
# Specter Vision Backend Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18

# SSE Support
sse-starlette>=2.2.0

# Google Gemini API
google-genai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
